Implementando as resposta do json

+++ src/main/kotlin/com/company/confinance/controller/ObjectiveController.kt
@@ -1,8 +1,10 @@
 package com.company.confinance.controller
 
 import com.company.confinance.model.entity.ObjectiveModel
+import com.company.confinance.model.response.CustomResponse
 import com.company.confinance.repository.ObjectiveRepository
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
@@ -25,31 +27,53 @@ class ObjectiveController {
     @PostMapping
     fun createObjective(
         @RequestBody objective : ObjectiveModel
-    ): ObjectiveModel {
-        return  repository.save(objective)
+    ): ResponseEntity <Any> {
+         return try {
+             ResponseEntity.status(HttpStatus.CREATED).body(repository.save(objective))
+         } catch (ex: Exception) {
+             ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Erro ao criar objetivo")
+         }
     }
 
     @DeleteMapping("/{id}")
-    fun delete(@PathVariable id: Long) {
-        return repository.deleteById(id)
+    fun delete(@PathVariable(value = "id") id: Long
+    ): ResponseEntity <Any> {
+        val existingObjective = repository.findById(id)
+        return if (id <= 0) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(
+                CustomResponse("ID do objectivo inválido", HttpStatus.BAD_REQUEST.value())
+            )
+        } else if (existingObjective.isPresent) {
+            repository.deleteById(id)
+            ResponseEntity.ok().body(CustomResponse
+                ("Objetivo Deletado com sucesso", HttpStatus.OK.value()))
+        } else {
+            ResponseEntity.status(HttpStatus.NOT_FOUND).body(
+                CustomResponse("Não foi possível deletar o objetivo, verifique o id informado.", HttpStatus.NOT_FOUND.value())
+            )
+
+        }
     }
 
     @PutMapping("/{id}")
     fun updateObjectiveById(
-        @PathVariable ("id") id: Long,
+        @PathVariable (value = "id") id: Long,
         @RequestBody objective: ObjectiveModel
-    ): ResponseEntity<ObjectiveModel> {
-        val existingObjective = repository.findById(id)
-        if (existingObjective.isPresent) {
-            val updateObjective = existingObjective.get()
-            updateObjective.value = objective.value
-            updateObjective.description = objective.description
-            updateObjective.date = objective.date
-            repository.save(updateObjective)
-            return ResponseEntity.ok(updateObjective)
+    ): ResponseEntity<Any> {
+        return if (repository.existsById(id)) {
+            val existingObjective = repository.findById(id).get()
+            existingObjective.value = objective.value
+            existingObjective.description = objective.description
+            existingObjective.date = objective.date
+            repository.save(existingObjective)
+            ResponseEntity.ok(existingObjective)
+        } else {
+            ResponseEntity.status(HttpStatus.NOT_FOUND).body(
+                CustomResponse("Não foi possível atualizar o objetivo, por favor verefique o id fornecido.", HttpStatus.NOT_FOUND.value())
+            )
         }
-        return ResponseEntity.notFound().build()
     }
+
 
 
     @GetMapping("/{id}")
